+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Changelog
 --------------
 
 - **Enhancement**: Suggest just annotating with `@ReadOnlyComposable` if a function is only `@Composable` to access composition locals.
+- **Fix**: Fix false positive in `ComposeRedundantComposable` when invoking composable function parameters in the body.
+
+Special thanks to [@UniqueEVE](https://github.com/UniqueEVE) for contributing to this release!
 
 1.5.2
 -----